requirements.txt
asgiref==3.7.2
attrs==22.1.0
dj-database-url==1.2.0
Django==4.2.3
django-bootstrap4==23.2
django-crispy-forms==2.0
django-stdimage==6.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
dj-static==0.0.6